Update:
Fixed minor layout issue in PondInfoActivity
PartnerPartnerActivity renamed to PartnerLogActivity
PartnerLogActivity can only be accessed by a Partner Account
PartnerLogActivity can be accessed via HomeActivity -> Partners will not be transferred to PondInfoActivity without checking in.
Partners can time in and time out

NOTES:
We need to modify buttomnavigation for PARTNERS. When HOME is clicked, we should only transfer them to PartnerLogActivity, and not to HomeActivity. --- Currently implemented in PondInfoActivity ONLY for testing purposes.