Update:
Added graph for temp forecast and weather forecast in PondInfoActivity
Tapping on the cardview of the "reads" in PondInfoActivity shows the graph/info
Fixed add pond. May now add depth, width, and length that are less that 0 or have decimals

Added a couple of layouts...

Possible bugs:
If width and length have decimals, adding might cause an issue.